api/requirements.txt
fastapi
uvicorn
pydantic
nfl-data-py
openai
requests
python-dotenv
pandas==2.2.3
